fixed security config
