+++ src/main/resources/application.properties
@@ -25,6 +25,4 @@ spring.datasource.password=tcgxgl
 #drop n create table again, good for testing, comment this in production
 spring.jpa.hibernate.ddl-auto=validate
 
-autoReconnect=true&useSSL=false
-
 ##server.servlet.context-path=/api
